requirements.txt
boto3==1.34.0
keras==3.0.1
numpy==1.25.0
Pillow==10.1.0
streamlit==1.27.2
streamlit_image_comparison==0.0.4
tensorflow==2.15.0.post1
toml==0.10.2
